pyproject.toml
[tool.poetry]
name = "pvz-svc"
version = "0.1.0"
description = ""
authors = ["gal16v8d"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.109.1"
pymongo = "^4.13.0"
uvicorn = "^0.25.0"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.7"
pytest = "^8.2.0"
mongomock = "^4.1.2"
httpx = "^0.27.0"
black = "^24.4.2"
mypy = "^1.10.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
